added error control

+++ Kernel/include/semaphores.h
@@ -7,7 +7,7 @@
 
 void sem_post(uint8_t sem);
 void sem_wait(uint8_t sem);
-uint8_t sem_open(const char * name, uint64_t value);
+int64_t sem_open(const char * name, uint64_t value);
 void sem_close(uint8_t sem);
 uint8_t sem_isValid(uint8_t sem);
